server/requirements.txt
fastapi
uvicorn
python-multipart
python-socketio
aiofiles
cloudinary
pymongo
dnspython
python-dotenv
